+++ asdf/extension/__init__.py
@@ -7,6 +7,7 @@ from ._manager import ExtensionManager, get_cached_extension_manager
 from ._manifest import ManifestExtension
 from ._tag import TagDefinition
 from ._converter import Converter, ConverterProxy
+from ._compressor import Compressor
 from ._legacy import (
     AsdfExtension,
     AsdfExtensionList,
@@ -27,6 +28,7 @@ __all__ = [
     "TagDefinition",
     "Converter",
     "ConverterProxy",
+    "Compressor",
     # Legacy API
     "AsdfExtension",
     "AsdfExtensionList",