+++ asdf/tests/test_compression.py
@@ -1,5 +1,6 @@
 import io
 import os
+import lzma
 
 import numpy as np
 
@@ -8,6 +9,8 @@ import pytest
 import asdf
 from asdf import compression
 from asdf import generic_io
+from asdf import config_context
+from asdf.extension import Extension, Compressor
 
 from ..tests import helpers
 
@@ -33,10 +36,12 @@ def _get_sparse_tree():
 
 def _roundtrip(tmpdir, tree, compression=None,
                write_options={}, read_options={}):
+    write_options = write_options.copy()
+    write_options.update(all_array_compression=compression)
+
     tmpfile = os.path.join(str(tmpdir), 'test.asdf')
 
     ff = asdf.AsdfFile(tree)
-    ff.set_array_compression(tree['science_data'], compression)
     ff.write_to(tmpfile, **write_options)
 
     with asdf.open(tmpfile, mode="rw") as ff:
@@ -49,7 +54,6 @@ def _roundtrip(tmpdir, tree, compression=None,
     buff = io.BytesIO()
 
     ff = asdf.AsdfFile(tree)
-    ff.set_array_compression(tree['science_data'], compression)
     ff.write_to(buff, **write_options)
 
     buff.seek(0)
@@ -60,14 +64,13 @@ def _roundtrip(tmpdir, tree, compression=None,
     buff = io.BytesIO()
 
     ff = asdf.AsdfFile(tree)
-    ff.set_array_compression(tree['science_data'], compression)
     ff.write_to(generic_io.OutputStream(buff), **write_options)
 
     buff.seek(0)
     with asdf.open(generic_io.InputStream(buff), **read_options) as ff:
         helpers.assert_tree_match(tree, ff.tree)
 
-    return ff
+    return tmpfile
 
 
 def test_invalid_compression():
@@ -76,9 +79,7 @@ def test_invalid_compression():
     with pytest.raises(ValueError):
         ff.set_array_compression(tree['science_data'], 'foo')
     with pytest.raises(ValueError):
-        compression._get_decoder('foo')
-    with pytest.raises(ValueError):
-        compression._get_encoder('foo')
+        compression._get_compressor('foo')
 
 
 def test_get_compressed_size():
@@ -181,10 +182,60 @@ def test_set_array_compression(tmpdir):
 
     tree = dict(zlib_data=zlib_data, bzp2_data=bzp2_data)
     with asdf.AsdfFile(tree) as af_out:
-        af_out.set_array_compression(zlib_data, 'zlib')
-        af_out.set_array_compression(bzp2_data, 'bzp2')
+        af_out.set_array_compression(zlib_data, 'zlib', level=1)
+        af_out.set_array_compression(bzp2_data, 'bzp2', compresslevel=9000)
+        with pytest.raises(ValueError):
+            af_out.write_to(tmpfile)
+        af_out.set_array_compression(bzp2_data, 'bzp2', compresslevel=9)
         af_out.write_to(tmpfile)
 
     with asdf.open(tmpfile) as af_in:
         assert af_in.get_array_compression(af_in.tree['zlib_data']) == 'zlib'
         assert af_in.get_array_compression(af_in.tree['bzp2_data']) == 'bzp2'
+
+
+class LzmaCompressor(Compressor):
+    def compress(self, data, **kwargs):
+        comp = lzma.compress(data, **kwargs)
+        yield comp
+
+    def decompress(self, blocks, out, **kwargs):
+        decompressor = lzma.LZMADecompressor(**kwargs)
+        i = 0
+        for block in blocks:
+            decomp = decompressor.decompress(block)
+            out[i:i+len(decomp)] = decomp
+            i += len(decomp)
+        return i
+
+    @property
+    def label(self):
+        return b'lzma'
+
+class LzmaExtension(Extension):
+    @property
+    def extension_uri(self):
+        return "asdf://somewhere.org/extensions/lzma-1.0"
+
+    @property
+    def compressors(self):
+        return [LzmaCompressor()]
+
+def test_compression_with_extension(tmpdir):
+    tree = _get_large_tree()
+
+    with config_context() as config:
+        config.add_extension(LzmaExtension())
+
+        with pytest.raises(lzma.LZMAError):
+            _roundtrip(tmpdir, tree, 'lzma',
+                        write_options=dict(compression_kwargs={'preset':9000}))
+        fn = _roundtrip(tmpdir, tree, 'lzma',
+                       write_options=dict(compression_kwargs={'preset':6}))
+
+        hist = {'extension_class': 'asdf.tests.test_compression.LzmaExtension',
+                'extension_uri': 'asdf://somewhere.org/extensions/lzma-1.0',
+                'supported_compression': ['lzma']}
+
+        with asdf.open(fn) as af:
+            assert hist in af['history']['extensions']