+++ asdf/tests/test_file_format.py
@@ -127,8 +127,7 @@ def test_invalid_source(small_tree):
             ff2.blocks.get_block(2)
 
         with pytest.raises(IOError):
-            # ff2.blocks.get_block("http://127.0.0.1/")
-            ff2.blocks.get_block("http://0.42.42.42/")
+            ff2.blocks.get_block("http://ABadUrl.verybad/test.asdf")
 
         with pytest.raises(TypeError):
             ff2.blocks.get_block(42.0)