+++ CHANGES.rst
@@ -45,6 +45,8 @@
 
 - Remove the ``asdf.test`` method and ``asdf.__githash__`` attribute. [#943]
 
+- Add support for custom compression via extensions. [#931]
+
 - Remove unnecessary ``.tree`` from search result paths. [#954]
 
 - Drop support for bugs in older operating systems and Python versions. [#955]
@@ -60,6 +62,8 @@
 - Copy array views when the base array is non-contiguous. [#949]
 
 - Prohibit views over FITS arrays that change dtype. [#952]
+
+- Add support for HTTPS URLs and following redirects. [#971]
 
 2.7.3 (2021-02-25)
 ------------------